Split test cases into file per query
