+++ api/juno/tests/base.py
@@ -1,9 +1,6 @@
 import json
 from django.test import TestCase
 
-from juno.schema import schema
-from juno.api.models import Sample, Institution, User
-
 
 class GraphQLTestCase(TestCase):
     """
@@ -188,166 +185,3 @@ class APITestCase(GraphQLTestCase):
             }
             """,
         )
-
-
-class SamplesQueryTestCase(APITestCase):
-    def setUp(self):
-        # put something in the db
-        self.institution = Institution.objects.create(
-            name="Wellcome Sanger Institute",
-            country="United Kingdom",
-            latitude=52.083333,
-            longitude=0.183333,
-        )
-        self.sample = Sample.objects.create(
-            lane_id="31663_7#113",
-            sample_id="5903STDY8059170",
-            public_name="CUHK_GBS177WT_16",
-            serotype="Ia",
-            host_status="skin and soft-tissue infection",
-            submitting_institution=self.institution,
-        )
-
-    def make_and_validate_samples_query(self, subquery):
-        # call api
-        samples = self.make_and_validate_query("samples", subquery)
-
-        # check non-empty
-        self.validate_non_empty_list(samples)
-
-        # return samples
-        return samples
-
-    def test_has_field_lane_id(self):
-        samples = self.make_and_validate_samples_query("laneId")
-        self.validate_field(
-            samples[0], "laneId", expected_value=self.sample.lane_id,
-        )
-
-    def test_has_field_sample_id(self):
-        samples = self.make_and_validate_samples_query("sampleId")
-        self.validate_field(
-            samples[0], "sampleId", expected_value=self.sample.sample_id,
-        )
-
-    def test_has_field_public_name(self):
-        samples = self.make_and_validate_samples_query("publicName")
-        self.validate_field(
-            samples[0], "publicName", expected_value=self.sample.public_name,
-        )
-
-    def test_has_field_serotype(self):
-        samples = self.make_and_validate_samples_query("serotype")
-        self.validate_field(
-            samples[0], "serotype", expected_value="IA",
-        )
-
-    def test_has_field_host_status(self):
-        samples = self.make_and_validate_samples_query("hostStatus")
-        self.validate_field(
-            samples[0],
-            "hostStatus",
-            expected_value="SKIN_AND_SOFT_TISSUE_INFECTION",
-        )
-
-    def test_has_field_submitting_institution(self):
-        samples = self.make_and_validate_samples_query(
-            """
-            submittingInstitution {
-                name
-            }
-            """,
-        )
-        self.validate_non_empty_list(samples)
-        submitting_institution = self.validate_field(
-            samples[0], "submittingInstitution"
-        )
-        self.validate_field(
-            submitting_institution,
-            "name",
-            expected_value=self.institution.name,
-        )
-
-
-class InstitutionsQueryTestCase(APITestCase):
-    def setUp(self):
-        # put something in the db
-        self.institution = Institution.objects.create(
-            name="Wellcome Sanger Institute",
-            country="United Kingdom",
-            latitude=52.083333,
-            longitude=0.183333,
-        )
-
-    def make_and_validate_institutions_query(self, subquery):
-        # call api
-        institutions = self.make_and_validate_query("institutions", subquery)
-
-        # check non-empty
-        self.validate_non_empty_list(institutions)
-
-        # return institutions
-        return institutions
-
-    def test_has_field_name(self):
-        institutions = self.make_and_validate_institutions_query("name")
-        self.validate_field(
-            institutions[0], "name", expected_value=self.institution.name,
-        )
-
-    def test_has_latitude(self):
-        institutions = self.make_and_validate_institutions_query("latitude")
-        self.validate_field(
-            institutions[0],
-            "latitude",
-            expected_value=self.institution.latitude,
-        )
-
-    def test_has_longitude(self):
-        institutions = self.make_and_validate_institutions_query("longitude")
-        self.validate_field(
-            institutions[0],
-            "longitude",
-            expected_value=self.institution.longitude,
-        )
-
-    def test_has_field_country(self):
-        institutions = self.make_and_validate_institutions_query("country")
-        self.validate_field(
-            institutions[0],
-            "country",
-            expected_value=self.institution.country,
-        )
-
-
-class MeQueryTestCase(APITestCase):
-    def setUp(self):
-        # put something in the db
-        self.user = User.objects.create(
-            email="bob@bob.com", first_name="Bob", last_name="Bobbity",
-        )
-        self.user.set_password("bobsicle")
-        self.user.save()
-
-    def make_me_query(self, subquery):
-        # call api
-        return self.query(
-            """
-            query {
-                me {
-                    %s
-                }
-            }
-            """
-            % subquery,
-        )
-
-    def test_has_field_email_when_logged_in(self):
-        # auth
-        self.login(self.user.email, self.user.password)
-
-        # query
-        response = self.make_me_query("email")
-        data = self.validate_successful(response)
-        me = self.validate_field(data, "me")
-        self.validate_field(me, "email", expected_value=self.user.email)
